django-dockerized/requirements.txt
lxml
requests
beautifulsoup4
djangorestframework